\name{get.header.index}
\alias{get.header.index}
\title{# note: lines after the header may be comments with '
# read.table should therefore be called with (skip=header.index, comment.char=')}
\usage{
  get.header.index(filepath)
}
\description{
  # note: lines after the header may be comments with ' #
  read.table should therefore be called with
  (skip=header.index, comment.char=')
}
\keyword{internal}

