\docType{package}
\name{qiime.io-package}
\alias{qiime.io-package}
\title{Support for loading and interacting with QIIME files in R.}
\description{
  There are already several ecology and phylogenetic
  packages available in R, including adephylo, vegan, ade4,
  picante, ape, phangorn, phylobase, phyloseq and OTUbase
  packages. This package is intended to be a standard
  interface to QIIME output files, especially legacy QIIME
  formats. See the biom package for help importing files in
  biom-format.
}
\author{
  Dan Knights <dan.knights@gmail.com>, Paul J. McMurdie
  <mcmurdie@stanford.edu>, and the QIIME devel team.
}
\references{
  \url{http://qiime.github.io/qiime.io/}
}
\keyword{package}

